refactor(web): split gem-success page presentational components
Extracts Shell/Icon/Title/Lede/PurpleAmount/BalanceCard/TokenChip/
Actions/PrimaryLink/SecondaryLink/SparkleField + ANIMATIONS CSS into a
sibling _components.tsx so the page itself drops from 507 to 115 lines.
Unblocks downstream ARC-710 commits that were tripping the
check-file-length pre-commit hook on this file.

No behavior change.

(ARC-710)
